nix: migrate away from cargo2nix as it's not needed at the moment
There doesn't seem to be a recent release and there isn't support for
1.86.0 in the latest release. After lots of messing around I've ended up
using the oxalica/rust-overlay directly for the devShell which gives a lot
more control over the rust version and toolchain as far as I can tell.

This change removes the nix package build and the app definition. However
these can be added back in later when I work out how. It looks like there
is builtin support for Rust builds with nix, so maybe a better route is to
use the builtin functions. That's for fun for another day though.

Signed-off-by: David James McCorrie <djmccorrie@gmail.com>